Neu Unit test.cpp and header (ausgelagert von Player) und map.h(to "simulate" the map.h of develop branch for CPU players)
