+++ PlayerEngine/PlayerEngine.vcxproj
@@ -129,10 +129,13 @@
   <ItemGroup>
     <ClCompile Include="..\CPU_opponent.cpp" />
     <ClCompile Include="..\Player.cpp" />
+    <ClCompile Include="..\Unit_test.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\CPU_opponent.h" />
+    <ClInclude Include="..\map.h" />
     <ClInclude Include="..\Player.h" />
+    <ClInclude Include="..\Unit_test.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
