+++ PlayerEngine/PlayerEngine.vcxproj.filters
@@ -21,12 +21,21 @@
     <ClCompile Include="..\CPU_opponent.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Unit_test.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\Player.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\CPU_opponent.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\Unit_test.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\map.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
