+++ Unit_test.h
@@ -1,0 +1,5 @@
+#pragma once
+class Map; // temp
+
+struct Property; // temp
+std::vector<Property> getTempPropertiesForPlayer(int playerID); //  temp
